[LIBK]: Fix Critical Mistake with u32 type
Interrupts haven't been working on the x86-64 Arch for over a week. I think this commit should fix that problem.

+++ lib/include/libk/types.h
@@ -2,8 +2,8 @@
 #define LIBK_TYPES_H
 
 typedef unsigned long long size_t;
-typedef unsigned long long u64;
-typedef unsigned long u32;
+typedef unsigned long u64;
+typedef unsigned int u32;
 typedef unsigned short u16;
 typedef unsigned char u8;
 #define NULL ((void*)0)
